Grove: untested cleanup
